Users controller

+++ VetManage/VetManage.Web/Controllers/OwnersController.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Vereyon.Web;
 using VetManage.Web.Data.Entities;
 using VetManage.Web.Data.Repositories;
 using VetManage.Web.Helpers;
@@ -25,6 +26,7 @@ namespace VetManage.Web.Controllers
         private readonly IUserHelper _userHelper;
         private readonly IMailHelper _mailHelper;
         private readonly IMessageHelper _messageHelper;
+        private readonly IFlashMessage _flashMessage;
 
         public OwnersController(
             IOwnerRepository ownerRepository,
@@ -33,7 +35,8 @@ namespace VetManage.Web.Controllers
             IBlobHelper blobHelper,
             IUserHelper userHelper,
             IMailHelper mailHelper,
-            IMessageHelper messageHelper)
+            IMessageHelper messageHelper,
+            IFlashMessage flashMessage)
         {
             _ownerRepository = ownerRepository;
             _messageBoxRepository = messageBoxRepository;
@@ -42,6 +45,7 @@ namespace VetManage.Web.Controllers
             _userHelper = userHelper;
             _mailHelper = mailHelper;
             _messageHelper = messageHelper;
+            _flashMessage = flashMessage;
         }
 
         // GET: Owners
@@ -83,7 +87,15 @@ namespace VetManage.Web.Controllers
         // GET: Vets/Create
         public IActionResult Create()
         {
-            return View();
+            var model = new RegisterOwnerViewModel
+            {
+                OwnerViewModel = new OwnerViewModel
+                {
+                    DateOfBirth = DateTime.Now,
+                },
+            };
+
+            return View(model);
         }
 
         // POST: Owners/Create
@@ -109,7 +121,7 @@ namespace VetManage.Web.Controllers
                         // Convert Owner
                         var owner = _converterHelper.ToOwner(model.OwnerViewModel, true, imageId);
 
-                        user = _converterHelper.ToUser(owner, new User());
+                        user = _converterHelper.ToUser(owner, new User(), "owners");
 
                         user.Email = model.Username;
                         user.UserName = model.Username;
@@ -122,9 +134,7 @@ namespace VetManage.Web.Controllers
 
                         if (result != IdentityResult.Success)
                         {
-                            ModelState.AddModelError(string.Empty, "The User could not be created.");
-
-                            return View(model);
+                            throw new Exception("The user could not be created, please try again.");
                         }
 
                         // Add role or roles to user
@@ -143,24 +153,24 @@ namespace VetManage.Web.Controllers
 
                         if (response.IsSuccess)
                         {
-                            ViewBag.Message = "Confirmation email has been sent";
-                            return RedirectToAction(nameof(Index));
+                            _flashMessage.Confirmation("Owner has been created and confirmation email has been sent to user.");
                         }
-                        ModelState.AddModelError(string.Empty, "Failed to Login");
 
-                        return RedirectToAction(nameof(Index));
-                        // TODO: Owner could not be created
-                    } else
-                    {
-                        // TODO: Owner already exists
+                        model.OwnerViewModel.ImageId = imageId;
+
+                        return View(model);
                     }
+
+                    _flashMessage.Danger("That email is already being used by another user.");
+
+                    return View(model);
                 }
                 catch (Exception ex)
                 {
-
+                    _flashMessage.Danger(ex.Message);
                 }
             }
-            return RedirectToAction(nameof(Index));
+            return View(model);
         }
 
         // GET: Owners/Edit/5
@@ -194,14 +204,14 @@ namespace VetManage.Web.Controllers
             {
                 try
                 {
-                    Guid imageId = Guid.Empty;
+                    Guid imageId = model.ImageId;
 
                     if (model.ImageFile != null && model.ImageFile.Length > 0)
                     {
                         imageId = await _blobHelper.UploadBlobAsync(model.ImageFile, "owners");
                     }
 
-                    // Get the user the user chose from the dropdown with the id
+                    // Get the user related with the owner
                     var user = await _userHelper.GetUserByIdAsync(model.UserId);
 
                     if(user == null)
@@ -211,7 +221,9 @@ namespace VetManage.Web.Controllers
 
                     var owner = _converterHelper.ToOwner(model, false, imageId);
 
-                    user = _converterHelper.ToUser(owner, user);
+                    user = _converterHelper.ToUser(owner, user, "owners");
+
+                    model.ImageId = imageId;
 
                     // Update the user 
                     var response = await _userHelper.UpdateUserAsync(user);
@@ -220,11 +232,20 @@ namespace VetManage.Web.Controllers
                     {
                         model.User = user;
 
+                        owner.UserId = user.Id;
+
                         await _ownerRepository.UpdateAsync(owner);
+
+                        _flashMessage.Confirmation("Owner has been updated.");
+
+                        return View(model);
                     }
-                    // TODO: Owner could not be updated
+
+                    _flashMessage.Danger("An error ocurred whilst tryng to update the owner, please try again.");
+
+                    return View(model);
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (Exception ex)
                 {
                     if (!await _ownerRepository.ExistsAsync(model.Id))
                     {
@@ -232,10 +253,9 @@ namespace VetManage.Web.Controllers
                     }
                     else
                     {
-                        throw;
+                        _flashMessage.Danger(ex.Message);
                     }
                 }
-                return RedirectToAction(nameof(Index));
             }
             return View(model);
         }
@@ -244,18 +264,25 @@ namespace VetManage.Web.Controllers
         {
             if (id == null)
             {
-                // vet not found
+                // owner not found
                 return new NotFoundViewResult("OwnerNotFound");
             }
 
             var owner = await _ownerRepository.GetWithUserByIdAsync(id.Value);
 
+            if(owner == null || owner.User == null)
+            {
+                // owner not found
+                return new NotFoundViewResult("OwnerNotFound");
+            }
+
             try
             {
                 var user = await _userHelper.GetUserByIdAsync(owner.User.Id);
 
-                await _ownerRepository.DeleteAsync(owner);
                 await _userHelper.DeleteUserAsync(user);
+
+                _flashMessage.Confirmation("Owner deleted successfully");
 
                 return RedirectToAction(nameof(Index));
 
@@ -304,7 +331,7 @@ namespace VetManage.Web.Controllers
 
             Response response = _mailHelper.SendEmail(
                 model.Username,
-                "Email Confirmation",
+                "Activate Account",
                 $"<h1>Email Confirmation</h1>" +
                 $"To activate your account please click the link and set up a new password:</br></br><a href = \"{tokenLink}\">Confirm Account</a>");
